debug

+++ gloo/connection.h
@@ -2239,10 +2239,10 @@ public:
 
                     if (out_off == ELICIT_OFFSET) {
                         /* Single Elicit packet */
-                        send_message[sent].setMessageHeader(pn, out_off, pkg_difference, (Packet_num_len)out_len, static_cast<Importance_len>(Channel::Important), out_blocks, static_cast<Status_len>(pkt_status), static_cast<uint8_t>(Type::ElicitAck));
+                        send_message[sent].setMessageHeader(pn, out_off, pkg_difference, (Packet_num_len)out_len, static_cast<Importance_len>(Channel::Important), out_blocks, static_cast<Status_len>(pkt_status), Type::ElicitAck);
                         recovery.on_packet_sent(out_len);
 
-                        connection_map.push(out_off, pkg_difference, out_blocks, static_cast<uint8_t>(PktStatus::Important_reliable), static_cast<uint8_t>(Type::ElicitAck));
+                        connection_map.push(out_off, pkg_difference, out_blocks, static_cast<uint8_t>(PktStatus::Important_reliable), Type::ElicitAck);
                     } else {
                         bool complete_flag = pkt_status == PktStatus::Unimportant_partialreliable;
                         send_message[sent].setMessageHeader(pn, out_off, pkg_difference, (Packet_num_len)out_len, static_cast<Importance_len>(Channel::Unimportant), out_blocks, (uint8_t)complete_flag);
